Move handleLoadItem handler in product
